refactor: 동물병원 패키지 네이밍 변경
